Changed the reduction calculation inside LMR.  The move count part of the calculation orignally used all integers, causing that part to always equal zero (because we divided 2/3 and multiplied that by the base value).  One of the numbers in the calculation is now a floating point number and the end result is cast back to an integer.

+++ Anduril.cpp
@@ -597,7 +597,7 @@ int Anduril::negamax(libchess::Position &board, int depth, int alpha, int beta, 
             || (!board.is_capture_move(move)
             && !check))) {
             // find our reduction
-            int reduction = reductions[depth] + reductions[moveCounter] * 2 / 3;
+            int reduction = reductions[depth] + (int)(reductions[moveCounter] * 2 / 3.0);
 
             // decrease if position is not likely to fail low
             if (PvNode && !likelyFailLow) {
